Read API Key from ENV variable or system property #10
The API key can be configured either in the VERSIONEYE_API_KEY environment variable or the versioneye.api.key system property to override settings from the property files/SBT configuration. The API key can be set in the environment variable and the system property but the plugin will fail if the value is different.